Move succeed(with:)/fail(with:) to TaskProtocol
